feature: 资源视图详情优化和yaml功能
