+++ bcs-app/frontend/src/views/dashboard/configuration/config-detail.css
@@ -1,0 +1,46 @@
+.detail {
+    /deep/ .bk-tab-label-item {
+        background-color: #FAFBFD;
+        border-bottom: 1px solid #dcdee5;
+        line-height: 39px !important;
+        height: 39px;
+        &.active {
+            border-bottom: none;
+        }
+    }
+    /deep/ .bk-tab-label-wrapper {
+        overflow: unset !important;
+    }
+    &-title {
+        margin-bottom: 10px;
+        font-size: 14px;
+        color: #333948;
+    }
+    &-content {
+        font-size: 14px;
+        margin-bottom: 20px;
+    }
+    &-table {
+        width: 100%;
+        border: 1px solid #dfe0e5;
+        border-radius: 3px;
+        color: #737987;
+        th {
+            text-align: left;
+            padding-left: 15px;
+            height: 32px;
+        }
+        tr {
+            height: 32px;
+            td {
+                padding-left: 15px;
+            }
+            td:nth-child(1) {
+                width: 215px;
+            }
+            &:nth-child(even) {
+                background-color: #F7F8FA;
+            }
+        }
+    }
+}
